SECURITY: The JWT signing key was previously stored in appsettings.json and committed to version control. Any party with repository access — past or present — must be considered to have had access to this key. Rotate immediately and purge from git history before deploying.  WHY these changes were necessary: - A hardcoded JWT secret in source control allows token forgery for   any user and any role. In a banking context this is a critical   vulnerability — an attacker could impersonate admins, approve   transfers, or access any account without valid credentials. - appsettings.json is committed to git. Deleting a secret from the   file does not remove it from history. The exposure surface includes   every developer machine, CI runner, and git hosting provider that   ever cloned the repository.  Changes: - Remove Jwt:Key from appsettings.json and appsettings.Development.json;   non-secret config (Issuer, Audience, ExpireInMinutes) is retained - Supply key via dotnet user-secrets in development and environment   variables in production — neither is committed to version control - Add fail-fast guard in Program.cs: application refuses to start if   Jwt:Key is absent or shorter than 32 bytes (256-bit minimum for   HMAC-SHA256); a misconfigured server must not silently start - Mirror the same key-length validation in TokenService as a defence-   in-depth measure at the point of token generation - Add .gitignore rules blocking secrets.json, appsettings.Local.json,   appsettings.Production.json, and *.env from being committed  Security impact: - JWT secret is no longer recoverable from repository history once   git history is purged and a new key is rotated into production - All existing sessions are invalidated on deployment of the new key;   active refresh tokens should also be revoked via SQL before deploy - Key-length enforcement prevents accidental use of weak secrets in   any future environment

+++ BankApi/Controllers/Account/AccountsController.cs
@@ -147,7 +147,7 @@ namespace ApiBank.Controllers.Account
         }
 
         [HttpPut("Deposit")]
-        public async Task<IActionResult> Deposit(string accountNumber, double balance)
+        public async Task<IActionResult> Deposit(string accountNumber, decimal balance)
         {
             if (string.IsNullOrWhiteSpace(accountNumber))
                 return BadRequest(new { success = false, message = "Account number is required." });
@@ -172,7 +172,7 @@ namespace ApiBank.Controllers.Account
         }
 
         [HttpPut("Withdraw")]
-        public async Task<IActionResult> Withdraw(string accountNumber, double balance)
+        public async Task<IActionResult> Withdraw(string accountNumber, decimal balance)
         {
             if (string.IsNullOrWhiteSpace(accountNumber))
                 return BadRequest(new { success = false, message = "Account number is required." });
@@ -198,7 +198,7 @@ namespace ApiBank.Controllers.Account
 
 
         [HttpPut("Transfer")]
-        public async Task<IActionResult> TransferAmount(string senderId, string receiverId, double balance)
+        public async Task<IActionResult> TransferAmount(string senderId, string receiverId, decimal balance)
         {
             try
             {
